requirements.txt
allure-pytest==2.8.22
allure-python-commons==2.8.22
aniso8601==9.0.1
apipkg==1.5
Appium-Python-Client==1.0.1
arrow==0.17.0
asgiref==3.2.7
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2019.6.16
cffi==1.13.1
chardet==3.0.4
click==8.0.1
colorama==0.4.1
comm==0.0.1
configparser==5.0.1
contextlib2==0.6.0.post1
cryptography==2.8
dataclasses==0.8
ddt==1.4.1
Django==3.0.5
enum34==1.1.10
execnet==1.7.1
Faker==4.14.2
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.8
importlib-metadata==0.19
itsdangerous==2.0.1
Jinja2==3.0.1
join==0.1.1
jsane==0.1.2
lib==3.0.0
lxml==4.6.2
MarkupSafe==2.0.1
marshmallow==2.15.4
more-itertools==7.2.0
multipart==0.2.4
mysqlclient==2.0.1
namedlist==1.8
packaging==19.1
paramiko==2.6.0
pluggy==0.12.0
psycopg2==2.8.4
py==1.8.0
pycparser==2.19
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.2
pytest==5.0.1
pytest-forked==1.1.3
pytest-html==2.0.1
pytest-metadata==1.8.0
pytest-order==1.0.0
pytest-ordering==0.6
pytest-repeat==0.9.1
pytest-rerunfailures==8.0
pytest-xdist==1.31.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
repeat==0.1.3
requests==2.22.0
requests-toolbelt==0.9.1
schema==0.7.3
selenium==3.141.0
sign==0.0.2
six==1.12.0
soupsieve==2.2.1
SQLAlchemy==1.4.21
sqlparse==0.3.1
sshtunnel==0.1.5
text-unidecode==1.3
urllib3==1.25.3
views==0.3
wcwidth==0.1.7
Werkzeug==2.0.1
xlrd==1.2.0
zipp==0.5.2
